// src/app/shared/models/filtros/filtro-compra.model.ts
export interface FiltroCompraDTO {
    numCompra?: number;
    idInciso?: number;
    idUnidadEjecutora?: number;
    idUnidadCompra?: number;
    anioCompra?: number;
    nroAnioCompra?: string;
    idCompra?: number;
    estado?: any;
    tipoCompra?: string;
    nroItem?: number;
    codArticulo?: number;
    descripcionArticulo?: string;
    idUsuario?: string;
    idTipoCompra?: string;
}